+++ cli/commands/toolchain/lintoko.ts
@@ -18,6 +18,12 @@ export let getReleases = async ({ prerelease = false } = {}) => {
   return toolchainUtils.getReleases(repo, { prerelease });
 };
 
+export let getReleaseTags = async (
+  options: toolchainUtils.ReleaseTagOptions = {},
+) => {
+  return toolchainUtils.getReleaseTags(repo, options);
+};
+
 export let isCached = (version: string) => {
   let dir = path.join(cacheDir, version);
   return fs.existsSync(dir) && fs.existsSync(path.join(dir, "lintoko"));